+++ src/gui.js
@@ -67,7 +67,7 @@
 */
 
 /*global modules, Morph, SpriteMorph, SyntaxElementMorph, Color, Cloud, Audio,
-ListWatcherMorph, TextMorph, newCanvas, useBlurredShadows, Sound, Scene,
+ListWatcherMorph, TextMorph, newCanvas, useBlurredShadows, Sound, Scene, Note,
 StringMorph, Point, MenuMorph, morphicVersion, DialogBoxMorph, normalizeCanvas,
 ToggleButtonMorph, contains, ScrollFrameMorph, StageMorph, PushButtonMorph, sb,
 InputFieldMorph, FrameMorph, Process, nop, SnapSerializer, ListMorph, detect,
@@ -79,7 +79,7 @@ CommandBlockMorph, BooleanSlotMorph, RingReporterSlotMorph, ScriptFocusMorph,
 BlockLabelPlaceHolderMorph, SpeechBubbleMorph, XML_Element, WatcherMorph, WHITE,
 BlockRemovalDialogMorph,TableMorph, isSnapObject, isRetinaEnabled, SliderMorph,
 disableRetinaSupport, enableRetinaSupport, isRetinaSupported, MediaRecorder,
-Animation, BoxMorph, BlockEditorMorph, BlockDialogMorph, Note, ZERO, BLACK*/
+Animation, BoxMorph, BlockEditorMorph, BlockDialogMorph, Project, ZERO, BLACK*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
@@ -4710,12 +4710,43 @@ IDE_Morph.prototype.exportProject = function (name, plain) {
     // newWindow requests displaying the project in a new tab.
     var menu, str, dataPrefix;
 
+    if (this.scenes.length() > 1) { // +++
+        this.exportScenes(this.scenes.at(1).name, plain);
+        return;
+    }
+
     if (name) {
         this.setProjectName(name);
         dataPrefix = 'data:text/' + plain ? 'plain,' : 'xml,';
         try {
             menu = this.showMessage('Exporting');
             str = this.serializer.serialize(this.scene);
+            this.setURL('#open:' + dataPrefix + encodeURIComponent(str));
+            this.saveXMLAs(str, name);
+            menu.destroy();
+            this.recordSavedChanges();
+            this.showMessage('Exported!', 1);
+        } catch (err) {
+            if (Process.prototype.isCatchingErrors) {
+                this.showMessage('Export failed: ' + err);
+            } else {
+                throw err;
+            }
+        }
+    }
+};
+
+IDE_Morph.prototype.exportScenes = function (name, plain) { // +++
+    // experimental export of multi-scene project - under construction
+
+    var menu, str, dataPrefix;
+
+    if (name) {
+        this.setProjectName(name);
+        dataPrefix = 'data:text/' + plain ? 'plain,' : 'xml,';
+        try {
+            menu = this.showMessage('Exporting');
+            str = this.serializer.serialize(new Project(this.scenes));
             this.setURL('#open:' + dataPrefix + encodeURIComponent(str));
             this.saveXMLAs(str, name);
             menu.destroy();