+++ src/scenes.js
@@ -34,6 +34,7 @@
     the following list shows the order in which all constructors are
     defined. Use this list to locate code in this document:
 
+    Project
     Scene
 
     credits
@@ -50,6 +51,23 @@
 /*global modules, VariableFrame, StageMorph, SpriteMorph, Process, List*/
 
 modules.scenes = '2021-April-16';
+
+
+// Projecct /////////////////////////////////////////////////////////
+
+// I am a container for a set of one or more Snap! scenes,
+// the IDE operates on an instance of me
+
+// Project instance creation:
+
+function Project(scenes) {
+    this.name = '';
+    this.notes = '';
+    this.scenes = scenes || new List();
+
+    // for undeleting scenes - do not persist
+    this.trash = [];
+}
 
 // Scene /////////////////////////////////////////////////////////