+++ HISTORY.md
@@ -11,6 +11,7 @@
 ### 2021-04-16
 * scenes, store, gui: remember last edited sprite in a scene / project
 * scenes: removed Project class
+* scenes, store, gui: export multi-scene projects
 
 ### 2021-04-14
 * scenes: new Project class